Undo removal of oos bug (cmd_result_ignored)

+++ benchmarks/milestone_1/loop_once-useless_test/fixed.sh
@@ -7,6 +7,5 @@ for f in $DIR/*
 do
     cp $f $f.bak
    sed 's+$OLD+$NEW+g' $f.bak &gt; $f
-   [ -f "$f" ] && \
-   rm -f $f.bak
+   [ -f "$f" ] && rm -f $f.bak
 done
